feat: cloud logs maintenance view, carousel groups, admin permissions catalog + wide fixes
Batches together several in-flight threads from this session and prior
concurrent sessions:

- Cloud logs maintenance: listCloudLogEntries() data fetcher + CloudLogsListView,
  backing /admin/maintenance/cloud-logs (Google Cloud Logging, bounded single
  getEntries() call per Hobby-tier rule #6)
- Carousel groups: AdminCarouselGroupEditorView (name + active/draft status)
- Auth permissions catalog exported publicly (auth/permissions/constants.ts)
  for the new read-only /admin/permissions reference page
- SiteMark: new icon-only glyph primitive, distinct from SiteLogo's wordmark
- Templates feature removed: SellerTemplatesView + its route-map entries
- New audits: audit-media-filename-generators.mjs (keeps validateMediaFilename's
  regex table in sync with generateMediaFilename's dispatcher, W1-51 bug class)
- SSR/client default-filter divergence fixes on public listing pages (Root
  Cause #30): AuctionsListView (dateFrom/showEnded), ProductsIndexPageView +
  ArtStickersListView (stockQuantity/showSold), PreOrdersListView
  (stockQuantity/showClosed) — staleTime:Infinity was freezing SSR initialData
  that didn't mirror the client's own default-toggle filtering
- Auction detail page: hide the "min increment" hint once an auction has ended
- Tester checklist: grouped into admin-editable phases; seed data idempotency
  and Beyblade-catalog content fixes across ~17 seed files
- Various schema/type/component updates across auctions, orders, products,
  cart, promotions, events, admin, seller, and store features

+++ src/_internal/shared/features/cart/schema.ts
@@ -4,7 +4,7 @@ const cartItemBaseSchema = z.object({
   productId: z.string().min(1),
   productTitle: z.string(),
   productImage: z.string().default(""),
-  price: z.number().int().min(0),
+  price: z.number().min(0),
   currency: z.string().default("INR"),
   storeId: z.string(),
   storeName: z.string().default(""),
@@ -22,7 +22,7 @@ const cartItemBaseSchema = z.object({
   ]),
   isOffer: z.boolean().optional(),
   offerId: z.string().optional(),
-  lockedPrice: z.number().int().optional(),
+  lockedPrice: z.number().optional(),
 });
 
 export const addToCartSchema = cartItemBaseSchema;
